Fix wrong package move
