+++ app/src/main/java/com/starry/myne/ui/common/NoBooksAvailable.kt
@@ -1,4 +1,4 @@
-package com.starry.myne.ui.screens.common
+package com.starry.myne.ui.common
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Arrangement
