+++ app/src/main/java/com/starry/myne/ui/common/VerticalScrollbar.kt
@@ -15,7 +15,7 @@ limitations under the License.
 */
 
 
-package com.starry.myne.ui.screens.common
+package com.starry.myne.ui.common
 
 import androidx.compose.animation.core.animateFloatAsState
 import androidx.compose.animation.core.tween
