+++ app/src/main/java/com/starry/myne/ui/screens/categories/composables/CategoriesScreen.kt
@@ -47,7 +47,7 @@ import androidx.navigation.NavController
 import androidx.navigation.compose.rememberNavController
 import com.starry.myne.R
 import com.starry.myne.ui.navigation.Screens
-import com.starry.myne.ui.screens.common.CustomTopAppBar
+import com.starry.myne.ui.common.CustomTopAppBar
 import com.starry.myne.ui.screens.categories.viewmodels.CategoryViewModel
 import com.starry.myne.ui.theme.figeronaFont
 import java.util.Locale
