+++ app/src/main/java/com/starry/myne/ui/screens/categories/composables/CategoryDetailScreen.kt
@@ -58,10 +58,10 @@ import com.starry.myne.R
 import com.starry.myne.others.BookLanguage
 import com.starry.myne.others.NetworkObserver
 import com.starry.myne.ui.navigation.Screens
-import com.starry.myne.ui.screens.common.BookItemCard
-import com.starry.myne.ui.screens.common.CustomTopAppBar
-import com.starry.myne.ui.screens.common.NoBooksAvailable
-import com.starry.myne.ui.screens.common.ProgressDots
+import com.starry.myne.ui.common.BookItemCard
+import com.starry.myne.ui.common.CustomTopAppBar
+import com.starry.myne.ui.common.NoBooksAvailable
+import com.starry.myne.ui.common.ProgressDots
 import com.starry.myne.ui.screens.categories.viewmodels.CategoryViewModel
 import com.starry.myne.ui.screens.home.composables.LanguageItem
 import com.starry.myne.ui.screens.other.NetworkError
