+++ app/src/main/java/com/starry/myne/ui/screens/home/composables/BookDetailScreen.kt
@@ -95,7 +95,7 @@ import com.airbnb.lottie.compose.animateLottieCompositionAsState
 import com.airbnb.lottie.compose.rememberLottieComposition
 import com.starry.myne.MainActivity
 import com.starry.myne.R
-import com.starry.myne.ui.screens.common.ProgressDots
+import com.starry.myne.ui.common.ProgressDots
 import com.starry.myne.ui.screens.home.viewmodels.BookDetailViewModel
 import com.starry.myne.ui.screens.other.NetworkError
 import com.starry.myne.ui.screens.settings.viewmodels.ThemeMode
