+++ app/src/main/java/com/starry/myne/ui/screens/settings/composables/SettingsScreen.kt
@@ -51,7 +51,7 @@ import com.starry.myne.BuildConfig
 import com.starry.myne.MainActivity
 import com.starry.myne.R
 import com.starry.myne.ui.navigation.Screens
-import com.starry.myne.ui.screens.common.CustomTopAppBar
+import com.starry.myne.ui.common.CustomTopAppBar
 import com.starry.myne.ui.screens.settings.viewmodels.SettingsViewModel
 import com.starry.myne.ui.screens.settings.viewmodels.ThemeMode
 import com.starry.myne.ui.theme.figeronaFont
